% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cargaVulnerabilidades}
\alias{cargaVulnerabilidades}
\title{sdfsdfs}
\usage{
cargaVulnerabilidades(TotalVersionesURLPaises)
}
\arguments{
\item{TotalVersionesURLPaises}{}
}
\description{
Encontrar Lista de Versiones encontradas de URLs de paises para sacar vulns de cada version
}
